refactor(core): move test utilities from src to tests/helpers
Move FakeTimeProvider from src/taskdog_core/testing/ to tests/helpers/.
Test-only code should not be in the production package.

Changes:
- Move time_provider.py to tests/helpers/
- Move test_fake_time_provider.py to tests/helpers/
- Update imports to use tests.helpers
