--- packages/taskdog-core/src/taskdog_core/testing/__init__.py
@@ -1,9 +1,0 @@
-"""Testing utilities for taskdog-core.
-
-This module provides test doubles and utilities for testing
-time-dependent and other business logic.
-"""
-
-from taskdog_core.testing.time_provider import FakeTimeProvider
-
-__all__ = ["FakeTimeProvider"]
